+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This version includes various little changes in order to comply with the evolvin
 
 *  Per the section [*Clarification on error reporting*](https://geminiprotocol.net/news/2024_08_28.gmi#clarification-on-error-reporting), longer explanations of the error codes are now included in the returned error objects.
 *  Previous versions of this module would not parse a link line correctly if there was no spacing after the leading `=>`.  This was allowed the whole time, but [*Permit use of tabs in link lines and text lines*](https://geminiprotocol.net/news/2024_08_28.gmi#permit-use-of-tabs-in-link-lines-and-text-lines) made this a requirement.
+*  The built-in TOFU validation is now supposed to match a hostname and port, not just a hostname.  This introduces a new file format for the TOFU cache.  It will be upgraded automatically, but can also be changed via `Convert-PSGeminiKnownHostsFileFormat`, should you want to upgrade or downgrade the file manually for some reason.
 
 There are also small improvements to the packaging structure and documentation.